include TermInfo in the test project, stop using reflection to test it
